niri: Basic config rendering with Python
I'd do this with Nix, but it's not really suitable for the task.
If we want data from Nix we'll export it to JSON and parse it with the
render tool.
In extension this tool should have a couple of different ways to render
output and input configuration based on what monitors are plugged in and
stuff. Might be deprecated by more mature features within Niri
eventually.
This config thingy could also be extended to other applications that
might need dynamic reconfiguration (Can't think of any right now).