Disable timescaledb until nixos/nixpkgs#341421
pg_config is broken and it's bothersome to override because postgres
packages(plugins) are their own package scope so we can't simply
override some bullshit fix for timescale without overriding the entirety
of our selected postgres version. Which doesn't really matter because
we're enabling Python support in postgres and building it anyways, but
it's not right!