libvirt: Implement hook to (de)allocate hugepages for win11-4
You can't use shared memory along with transparent/anonymous hugepages,
and we use shared memory for virtfsd (and possibly Looking-Glass at some
point).

On win11-4 startup we do the following:
1. Drop caches
2. Compact memory
3. Allocate 8192 2M hugepages
4. Start VM

On win11-4 shutdown we do the following:
1. Deallocate all hugepages

We also add NixVirt here to (in the future) declare our gaming VM with
Nix rather than creating it with with Virt-Manager.