libvirt: Fix hooks not updating when edited
Hooks are stored in /var/lib/libvirt/hooks which is a mutable directory
we can't rely on nice things like environment.etc to make sure files are
what we want them to be.

Solution: Get all hook derivations and stick them in the restartTriggers
for libvirtd-config which is responsible for linking the hooks.