libvirt: Prematurely overoptimize CPU in gamingvm
Pin CPU threads according to the physical CPU topology of the shitbox
host machine. Pin QEMU and IO threads to different threads than what's
used for the guest.

My idea is that this will allow me to tune PipeWire and sound to lower
latencies since the scheduler won't compete with Windows hogging it's
cores.