Implement notification sound when pressing KP0
Only on the "macropad"
Uses a volume algorithm to lower the notification output volume the
louder the system configured output volume is. Notification sounds are
usually subject to a too low volume when on low volume and too high
on high volume but this solves it for all but the lowest volumes where you
wouldn't expect to hear anything anyways.